+++ .github/actions/parser/index.js
@@ -49,11 +49,20 @@ const PACKAGES = [
   {name: 'protractor', dirname: 'eyes-protractor', framework: 'protractor', sdk: true, aliases: ['@applitools/eyes-protractor'], dependencies: ['types', 'utils', 'api', 'core', 'vgc', 'test-utils']},
   {name: 'nightwatch', dirname: 'eyes-nightwatch', framework: 'nightwatch', sdk: true, aliases: ['nw', '@applitools/eyes-nightwatch'], dependencies: ['types', 'utils', 'api', 'core', 'vgc', 'test-utils']},
   {name: 'testcafe', dirname: 'eyes-testcafe', framework: 'testcafe', sdk: true, aliases: ['@applitools/eyes-testcafe'], dependencies: ['types', 'utils', 'api', 'core', 'vgc', 'test-utils']},
-  {name: 'browser-extension', dirname: 'eyes-browser-extension', sdk: true, aliases: ['extension', '@applitools/eyes-browser-extension'], dependencies: ['utils', 'core', 'vgc', 'spec-playwright', 'test-utils']},
+  {name: 'browser-extension', dirname: 'eyes-browser-extension', sdk: true, xvfb: true, aliases: ['extension', '@applitools/eyes-browser-extension'], dependencies: ['utils', 'core', 'vgc', 'spec-playwright', 'test-utils']},
   {name: 'cypress', dirname: 'eyes-cypress', framework: 'cypress', sdk: true, aliases: ['cy', '@applitools/eyes-cypress'], dependencies: ['logger', 'vgc', 'test-server']},
   {name: 'storybook', dirname: 'eyes-storybook', framework: 'storybook', sdk: true, aliases: ['@applitools/eyes-storybook'], dependencies: ['logger', 'core', 'vgc', 'spec-puppeteer', 'test-utils']},
   // #endregion
 ]
+
+const OS = {
+  linux: 'ubuntu-latest',
+  ubuntu: 'ubuntu-latest',
+  mac: 'macos-latest',
+  macos: 'macos-latest',
+  win: 'windows-2022',
+  windows: 'windows-2022',
+}
 
 const packageSettings = core.getInput('packages', {required: true})
 const allowVariations = core.getBooleanInput('allow-variations')
@@ -82,9 +91,13 @@ core.setOutput('packages', allowVariations ? Object.values(packages) : packages)
 
 function requestedPackages(packageSettings) {
   return packageSettings.split(/[\s,]+/).reduce((packages, packageSetting) => {
-    const [_, packageKey, releaseVersion = defaultReleaseVersion, frameworkVersion, frameworkProtocol]
-      = packageSetting.match(/^(.*?)(?::(patch|minor|major))?(?:@([\d.]+))?(?:\+(.+?))?$/i)
+    let [_, packageKey,  releaseVersion, frameworkVersion, frameworkProtocol, nodeVersion, jobOS, shortReleaseVersion, shortFrameworkVersion, shortFrameworkProtocol]
+      = packageSetting.match(/^(.*?)(?:\((?:version:(patch|minor|major);?)?(?:framework:([\d.]+);?)?(?:protocol:(.+?);?)?(?:node:([\d.]+);?)?(?:os:(linux|ubuntu|mac|macos|win|windows);?)?\))?(?::(patch|minor|major))?(?:@([\d.]+))?(?:\+(.+?))?$/i)
   
+    releaseVersion ??= shortReleaseVersion ?? defaultReleaseVersion
+    frameworkVersion ??= shortFrameworkVersion
+    frameworkProtocol ??= shortFrameworkProtocol
+
     const packageInfo = PACKAGES.find(({name, dirname, aliases}) => {
       return name === packageKey || dirname === packageKey || aliases.includes(packageKey)
     })
@@ -104,7 +117,7 @@ function requestedPackages(packageSettings) {
       }
     }
   
-    const appendix = Object.entries({release: releaseVersion, version: frameworkVersion, protocol: frameworkProtocol})
+    const appendix = Object.entries({release: releaseVersion, version: frameworkVersion, protocol: frameworkProtocol, os: jobOS, node: nodeVersion})
       .reduce((parts, [key, value]) => value ? [...parts, `${key}: ${value}`] : parts, [])
       .join('; ')
   
@@ -113,7 +126,10 @@ function requestedPackages(packageSettings) {
       name: packageInfo.name,
       dirname: packageInfo.dirname,
       sdk: packageInfo.sdk,
+      xvfb: packageInfo.xvfb,
       install: frameworkVersion ? `${packageInfo.framework}@${frameworkVersion}` : '',
+      os: OS[jobOS ?? 'linux'],
+      node: nodeVersion ?? 'lts/*',
       releaseVersion,
       env: {
         [`APPLITOOLS_${packageInfo.name.toUpperCase()}_MAJOR_VERSION`]: frameworkVersion,